pyproject.toml
[tool.poetry]
name = "electronics-trading-network"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "electronics_trading_network"}]

[tool.poetry.dependencies]
python = "^3.11"
djangorestframework = "^3.14.0"
psycopg2 = "^2.9.9"
pillow = "^10.2.0"
django-filter = "^23.5"
djangorestframework-simplejwt = "^5.3.1"
drf-yasg = "^1.21.7"
python-dotenv = "^1.0.0"
coverage = "^7.4.0"
django = "^4.2.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
